chore:split out NavigationMenu components
